# Conflicts:
#	core/src/com/cpe/magigo/Screens/PlayScreen.java

+++ core/src/com/cpe/magigo/Sprites/HP_Char.java
@@ -17,8 +17,8 @@ public class HP_Char extends Sprite {
     private Sprite HP_BG;
     private Texture BG = new Texture("HealtBar/source/health_bar.png");
     private TextureRegion BG_R;
-    public int Hp;
-    public HP_Char(int hp)
+    public float Hp;
+    public HP_Char(float hp)
     {
         this.Hp = hp;
         BG_R = new TextureRegion(BG);
@@ -29,14 +29,14 @@ public class HP_Char extends Sprite {
         HP_FG.setOrigin(0,0);
     }
 
-    public void update(int dt)
+    public void update(float dt)
     {
         this.Hp = dt;
         HP_BG.setPosition((MagiGO.V_WIDTH/2-270)/MagiGO.PPM, 30 / MagiGO.PPM);
         HP_BG.setScale(0.65f/MagiGO.PPM,0.5f/MagiGO.PPM);
-        HP_FG.setSize(this.Hp*(FG_R.getRegionWidth()/100),FG_R.getRegionHeight());
+        HP_FG.setSize((int)this.Hp*(FG_R.getRegionWidth()/100),FG_R.getRegionHeight());
         HP_FG.setPosition((MagiGO.V_WIDTH/2-210)/MagiGO.PPM, 30 / MagiGO.PPM);
-        HP_FG.setRegion(90,0,(this.Hp*(FG_R.getRegionWidth()/100)),FG_R.getRegionHeight());
+        HP_FG.setRegion(90,0,(int)(this.Hp*(FG_R.getRegionWidth()/100)),FG_R.getRegionHeight());
         HP_FG.setScale(0.65f/MagiGO.PPM,0.5f/MagiGO.PPM);
     }
